Fix Int/Float type mismatch in the stacked bar chart's tap handler
Inside pointerInput's tap handler, size.height is PointerInputScope's
IntSize.height (Int), unlike a Canvas draw scope's size.height (Float).
Assigning it straight to `var y` inferred an Int, which then failed to
compile against the Float segment heights used in range checks and the
running-total decrement. Explicitly convert to Float at the point of
assignment.

+++ app/src/main/java/com/azimulkabir/actua/ui/reports/ReportsScreen.kt
@@ -774,7 +774,7 @@ private fun StackedIntervalBars(points: List<ReportPoint>, hideDecimals: Boolean
                 detectTapGestures { tap ->
                     val i = (tap.x / size.width * points.size).toInt().coerceIn(0, points.size - 1)
                     val point = points[i]
-                    var y = size.height
+                    var y = size.height.toFloat()
                     var hit: ReportCategory? = null
                     for (segment in point.segments) {
                         val h = segment.spentCents.absoluteValue.toFloat() / maximum * size.height
